Make plugin platform independent
Refactor to primarily use Path's, and where necessary fallback to File.
These are filesystem independent abstractions on the locations of files
and directories, and as such, we don't need to worry as much about the
various path separators.

Introduce TranspileContext, an object containing details about a trans-
pilation fo a single source file. By initializing the transpilation
process with a list of these objects, it is easier to provide proper
feedback (logging which file is transpiled and to what target file).
Also, it reduces the amount of state required in the TargetFileWriter
and BabelTranspiler.

Increase logging when using verbose mode. When using verbose mode,
log most messages on the INFO level. Only logging that would really
clutter the output too much, or clearly has debugging purposes, use
logging level DEBUG.

Add lombok to reduce boilerplate. Introduce builder pattern for
increasedd readability.

Include an example project (as much to demonstrate the plugin's use, as
well as to test the plugin on various platforms)

This modification has been tested on MacOS, and Windows 10, running
JDK8 and Maven 3.6, both.